chat

+++ backend/src/main/java/com/example/social_media/dto/message/MessageDto.java
@@ -3,16 +3,16 @@ package com.example.social_media.dto.message;
 import java.time.Instant;
 
 public class MessageDto {
-    private Long id;
-    private Long chatId;
-    private Long senderId;
+    private Integer id;
+    private Integer chatId;
+    private Integer senderId;
     private String content;
     private Integer typeId;
     private Instant createdAt;
 
     public MessageDto() {}
 
-    public MessageDto(Long id, Long chatId, Long senderId, String content, Integer typeId, Instant createdAt) {
+    public MessageDto(Integer id, Integer chatId, Integer senderId, String content, Integer typeId, Instant createdAt) {
         this.id = id;
         this.chatId = chatId;
         this.senderId = senderId;
@@ -22,12 +22,12 @@ public class MessageDto {
     }
 
     // Getters and setters
-    public Long getId() { return id; }
-    public void setId(Long id) { this.id = id; }
-    public Long getChatId() { return chatId; }
-    public void setChatId(Long chatId) { this.chatId = chatId; }
-    public Long getSenderId() { return senderId; }
-    public void setSenderId(Long senderId) { this.senderId = senderId; }
+    public Integer getId() { return id; }
+    public void setId(Integer id) { this.id = id; }
+    public Integer getChatId() { return chatId; }
+    public void setChatId(Integer chatId) { this.chatId = chatId; }
+    public Integer getSenderId() { return senderId; }
+    public void setSenderId(Integer senderId) { this.senderId = senderId; }
     public String getContent() { return content; }
     public void setContent(String content) { this.content = content; }
     public Integer getTypeId() { return typeId; }
